Add asynchronous handlers for messages
Use asynchronous handlers for processing messages. A more complete and
fancy handling has to be done once we process messages that have some
functional dependency. This handling will be done in the future. Yet
this could only be a problem with corner cases, since the dependent
messages are sent by the node that creates and transforms `Silo`s and,
by design, this client manipulates it synchronously.

Also, it's worth noticing that the client business logic is executed
synchronously and it's not shared with other clients.

In case that those corner cases do exist, a good approach to solve them is
to flatMap on the last future that has transformed a given `Silo`. This
would require bookkeeping of these `Future`s for any `SiloRef` to which
a message is related.