Add hack to overcome pickler generation limitation
What happens when we want to generate `Pickler`s for subclasses that are
stored in a data structure as one of their super class? Surprise! Pickler
generation fails since it doesn't have enough type information at the
call site that queries that data structure (the type of the returned
value is the super class).

This hack is a workaround that uses path-dependent types to overcome
this limitation. Also, it forces the creation of picklers for generic
classes at the call site (see `Transformed`). In this case, this hack is
applied to `Response`s since we need to store the most recent sent message
to a client in a Map. While other approaches were available, like replicating
the logic and recreating the `Response` from the `Request`, this one is
the most neat and efficient one.